requirements.txt
annotated-types==0.6.0
anyio==3.7.1
black==24.1.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
cryptography==42.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.104.0
fastapi-debug-toolbar==0.5.0
fastapi_nextauth_jwt==1.1.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
humps==0.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
mangum==0.17.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.26.1
orjson==3.8.11
packaging==23.2
pandas==2.1.1
pathspec==0.12.1
platformdirs==4.1.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyinstrument==4.6.0
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0
requests==2.31.0
rsa==4.9
sentry-sdk==1.32.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.7.0
ulid==1.1
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
